+++ flask_security/datastore.py
@@ -322,6 +322,10 @@ class UserDatastore:
         :kwparam roles: list of roles to be added to user.
             Can be Role objects or strings
 
+        .. note::
+            No normalization is done on email - it is assumed the caller has already
+            done that.
+
         .. danger::
            Be aware that whatever `password` is passed in will
            be stored directly in the DB. Do NOT pass in a plaintext password!